Rename the project from Unified Catalog to Graviton
