+++ .github/ISSUE_TEMPLATE/feature-request.yml
@@ -1,6 +1,6 @@
 name: Feature Request
 title: "[FEATURE] "
-description: Suggest an idea for Unified Catalog
+description: Suggest an idea for Graviton
 labels: [ "kind:feature,priority:major" ]
 body:
