+++ core/src/main/java/com/datastrato/graviton/schema/Auditable.java
@@ -1,0 +1,11 @@
+package com.datastrato.graviton.schema;
+
+public interface Auditable {
+
+  /**
+   * Returns the audit information of the entity.
+   *
+   * @return AuditInfo
+   */
+  AuditInfo auditInfo();
+}
