--- core/src/main/java/com/datastrato/unified_catalog/schema/Auditable.java
@@ -1,11 +1,0 @@
-package com.datastrato.unified_catalog.schema;
-
-public interface Auditable {
-
-  /**
-   * Returns the audit information of the entity.
-   *
-   * @return AuditInfo
-   */
-  AuditInfo auditInfo();
-}
