+++ settings.gradle.kts
@@ -1,3 +1,3 @@
-rootProject.name = "Unified Catalog"
+rootProject.name = "Graviton"
 
 include("api", "core", "schema", "server")
